Face-Tracking-App/requirements.txt
absl-py==2.3.1
astunparse==1.6.3
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
decorator==5.2.1
facenet-pytorch==2.6.0
filelock==3.13.1
fire==0.7.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fsspec==2024.6.1
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.73.1
gunicorn==23.0.0
h5py==3.14.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.5.1
libclang==18.1.1
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2

mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
networkx==3.3
numpy==1.26.4
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.0
pillow==10.2.0
proglog==0.1.12
protobuf==5.29.5
Pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
rich==14.0.0
six==1.17.0
soupsieve==2.7
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2